Overrides and missing initializations

+++ src/games/chess.h
@@ -18,8 +18,8 @@ namespace chess {
 
 struct ChessBoard {
 
-  static const size_t boardSize = 8;
-  static const size_t boardDim = boardSize + 4;
+  static const int boardSize = 8;
+  static const int boardDim = boardSize + 4;
 
   std::array<uint8_t, boardDim * boardDim> board;
